+++ src/main/kotlin/io/snyk/plugin/snykcode/SnykCodeBulkFileListener.kt
@@ -7,10 +7,10 @@ import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task.Backgroundable
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.openapi.vfs.newvfs.events.VFileEvent
 import com.intellij.openapi.vfs.readText
 import io.snyk.plugin.SnykBulkFileListener
-import io.snyk.plugin.getPsiFile
 import io.snyk.plugin.getSnykCachedResults
 import io.snyk.plugin.getSnykTaskQueueService
 import io.snyk.plugin.isSnykCodeLSEnabled
@@ -79,9 +79,12 @@ class SnykCodeBulkFileListener : SnykBulkFileListener() {
     }
 
     override fun forwardEvents(events: MutableList<out VFileEvent>) {
+        val languageServerWrapper = LanguageServerWrapper.getInstance()
+
         if (!isSnykCodeLSEnabled()) return
-        LanguageServerWrapper.getInstance().ensureLanguageServerInitialized()
-        val languageServer = LanguageServerWrapper.getInstance().languageServer
+        if (!languageServerWrapper.isInitialized) return
+
+        val languageServer = languageServerWrapper.languageServer
         for (event in events) {
             if (event.file == null || !event.isFromSave) continue
             val file = event.file!!
@@ -102,10 +105,9 @@ class SnykCodeBulkFileListener : SnykBulkFileListener() {
         val cache = getSnykCachedResults(project)?.currentSnykCodeResultsLS ?: return
         filesAffected.forEach {
             cache.remove(it)
-            it.virtualFile.getPsiFile(project)?.let { psiFile ->
-                DaemonCodeAnalyzer.getInstance(project).restart(psiFile)
-            }
         }
+        VirtualFileManager.getInstance().asyncRefresh()
+        DaemonCodeAnalyzer.getInstance(project).restart()
     }
 
     private fun toSnykCodeFileSet(project: Project, virtualFiles: Set<VirtualFile>) =