+++ src/main/kotlin/snyk/common/AnnotatorCommon.kt
@@ -1,5 +1,6 @@
 package snyk.common
 
+import com.intellij.openapi.application.invokeLater
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
@@ -8,12 +9,17 @@ import io.snyk.plugin.events.SnykProductsOrSeverityListener
 import io.snyk.plugin.events.SnykSettingsListener
 import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.refreshAnnotationsForOpenFiles
+import snyk.common.lsp.LanguageServerWrapper
 
 object AnnotatorCommon {
     val logger = logger<AnnotatorCommon>()
 
     fun prepareAnnotate(psiFile: PsiFile?) {
         logger.debug("Preparing annotation for $psiFile")
+
+        // trigger LS initialization if not already done, we consciously don't check the result here
+        LanguageServerWrapper.getInstance().ensureLanguageServerInitialized()
+
         // todo: review later if any way to provide up-to-date context for CLI scans is available
         // force saving here will break some user's workflow: https://github.com/snyk/snyk-intellij-plugin/issues/324
     }
@@ -25,16 +31,21 @@ object AnnotatorCommon {
         logger.debug("Initializing annotations refresh listener")
         project.messageBus.connect()
             .subscribe(
-                SnykProductsOrSeverityListener.SNYK_ENABLEMENT_TOPIC, object : SnykProductsOrSeverityListener {
-                override fun enablementChanged() {
-                    refreshAnnotationsForOpenFiles(project)
+                SnykProductsOrSeverityListener.SNYK_ENABLEMENT_TOPIC,
+                object : SnykProductsOrSeverityListener {
+                    override fun enablementChanged() {
+                        invokeLater { refreshAnnotationsForOpenFiles(project) }
+                    }
                 }
-            })
+            )
         project.messageBus.connect()
-            .subscribe(SnykSettingsListener.SNYK_SETTINGS_TOPIC, object : SnykSettingsListener {
-                override fun settingsChanged() {
-                    refreshAnnotationsForOpenFiles(project)
+            .subscribe(
+                SnykSettingsListener.SNYK_SETTINGS_TOPIC,
+                object : SnykSettingsListener {
+                    override fun settingsChanged() {
+                        invokeLater { refreshAnnotationsForOpenFiles(project) }
+                    }
                 }
-            })
+            )
     }
 }