+++ src/main/kotlin/snyk/common/CustomEndpoints.kt
@@ -1,8 +1,8 @@
 package snyk.common
 
 import io.snyk.plugin.pluginSettings
-import io.snyk.plugin.suffixIfNot
 import io.snyk.plugin.prefixIfNot
+import io.snyk.plugin.suffixIfNot
 import java.net.URI
 import java.net.URISyntaxException